server/requirements.txt
blinker==1.7.0
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==3.0.1
